+++ source/config/OptionValidator.ts
@@ -1,8 +1,8 @@
-import { existsSync } from "node:fs";
+import { existsSync, watch } from "node:fs";
 import { Diagnostic, type DiagnosticOrigin } from "#diagnostic";
 import { Environment } from "#environment";
+import { Path } from "#path";
 import type { StoreService } from "#store";
-import { Watcher } from "#watch";
 import { OptionDiagnosticText } from "./OptionDiagnosticText.js";
 import { OptionUsageText } from "./OptionUsageText.js";
 import type { OptionBrand, OptionGroup } from "./enums.js";
@@ -57,7 +57,7 @@ export class OptionValidator {
           break;
         }
 
-        if (!Watcher.isSupported()) {
+        if (!this.#isWatchSupported()) {
           this.#onDiagnostic(Diagnostic.error(OptionDiagnosticText.watchIsNotAvailable(), origin));
         }
         break;
@@ -66,5 +66,20 @@ export class OptionValidator {
       default:
         break;
     }
+  }
+
+  #isWatchSupported(): boolean {
+    let isRecursiveWatchAvailable: boolean | undefined;
+
+    try {
+      const watcher = watch(Path.resolve("./"), { persistent: false, recursive: true });
+      watcher.close();
+
+      isRecursiveWatchAvailable = true;
+    } catch {
+      isRecursiveWatchAvailable = false;
+    }
+
+    return isRecursiveWatchAvailable;
   }
 }
